+++ triad-sample/app/src/main/java/com/nhaarman/triad/sample/notes/NotesContainer.java
@@ -1,6 +1,6 @@
 package com.nhaarman.triad.sample.notes;
 
-import com.nhaarman.triad.container.ScreenContainer;
+import com.nhaarman.triad.ScreenContainer;
 import com.nhaarman.triad.sample.notes.noteslist.NotesListContainer;
 import org.jetbrains.annotations.NotNull;
