+++ triad-sample/app/src/main/java/com/nhaarman/triad/sample/notes/noteslist/NotesListContainer.java
@@ -1,8 +1,7 @@
 package com.nhaarman.triad.sample.notes.noteslist;
 
-import com.nhaarman.triad.container.Container;
+import com.nhaarman.triad.Container;
 import com.nhaarman.triad.sample.Note;
-import java.util.Collection;
 import java.util.List;
 import org.jetbrains.annotations.NotNull;
