+++ triad-sample/app/src/main/java/com/nhaarman/triad/sample/notes/noteslist/NotesListPresenter.java
@@ -1,6 +1,6 @@
 package com.nhaarman.triad.sample.notes.noteslist;
 
-import com.nhaarman.triad.presenter.Presenter;
+import com.nhaarman.triad.Presenter;
 import com.nhaarman.triad.sample.Note;
 import com.nhaarman.triad.sample.NoteRepository;
 import java.util.List;
