+++ triad-tests/src/main/java/com/nhaarman/triad/tests/secondscreen/SecondScreenContainer.java
@@ -1,6 +1,6 @@
 package com.nhaarman.triad.tests.secondscreen;
 
-import com.nhaarman.triad.container.ScreenContainer;
+import com.nhaarman.triad.ScreenContainer;
 
 public interface SecondScreenContainer extends ScreenContainer<SecondScreenPresenter, SecondScreenContainer> {
 }
