+++ triad-tests/src/main/java/com/nhaarman/triad/tests/secondscreen/SecondScreenPresenter.java
@@ -1,6 +1,6 @@
 package com.nhaarman.triad.tests.secondscreen;
 
-import com.nhaarman.triad.presenter.ScreenPresenter;
+import com.nhaarman.triad.ScreenPresenter;
 
 public class SecondScreenPresenter extends ScreenPresenter<SecondScreenPresenter, SecondScreenContainer> {
 }
