+++ triad/src/main/java/com/nhaarman/triad/TriadActivity.java
@@ -2,8 +2,6 @@ package com.nhaarman.triad;
 
 import android.app.Activity;
 import android.os.Bundle;
-import com.nhaarman.triad.presenter.Presenter;
-import com.nhaarman.triad.screen.Screen;
 import flow.Flow;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
