+++ triad/src/main/java/com/nhaarman/triad/TriadDelegate.java
@@ -1,10 +1,6 @@
 package com.nhaarman.triad;
 
 import android.app.Activity;
-import com.nhaarman.triad.container.ScreenContainer;
-import com.nhaarman.triad.presenter.Presenter;
-import com.nhaarman.triad.presenter.ScreenPresenter;
-import com.nhaarman.triad.screen.Screen;
 import flow.Backstack;
 import flow.Flow;
 import java.util.AbstractList;
