+++ triad/src/main/java/com/nhaarman/triad/TriadPresenter.java
@@ -2,10 +2,6 @@ package com.nhaarman.triad;
 
 import android.view.View;
 import android.view.ViewGroup;
-import com.nhaarman.triad.container.ScreenContainer;
-import com.nhaarman.triad.presenter.Presenter;
-import com.nhaarman.triad.presenter.ScreenPresenter;
-import com.nhaarman.triad.screen.Screen;
 import flow.Flow;
 import flow.Flow.Direction;
 import java.util.Stack;
